move and rename MarkerIcon
